start creating a subclass for characters to choose from when meeting criteria

+++ src/game/actors/resources/ResourceValues.java
@@ -31,7 +31,7 @@ public abstract class ResourceValues {
         return regenValue;
     }
 
-    public void setRegenValue(int regenValue) {
+    public void setRegenValue(double regenValue) {
         this.regenValue = regenValue;
     }
